endpoint: Try immediate send before queing pending
sendWhenActive always queues the request for next time the pending queue
was flushed. This has the unfortunate side-effect of delaying sends
until next time the end device talks, even if the device is currently
active.

Instead, try to send the request immediately, and only fall back to the
pending queue if that failed or if something was already queued.