Add support for multiple data point values in Tuya payloads
Although most TuYa devices send only one data point value in a data
report/command, it is possible to include values for multiple data points in
such a payload.

Change the definitions of those receiving commands to decode all values
present.  Store them in the `dpValues` array.  Additionally, remove the `fn`
field: It is just the high byte of the length of the data buffer and isn't
used by converters.

**NOTE:** This breaks all existing Tuya converters using `dataResponse`,
`dataReport`, or `activeStatusReport`.  They need to be adapted to use the
first entry of `dpValues`