Clean up message queue after fastpoll send
Clean up everything that is left in the message queue after a fastpoll send.
Everything that is left after receiving a genPollCtrl checkin from a sleepy end device and enabling fastpoll must have failed to be sent due to other reasons than the sleeping device.
There is little value to fill up the message queue with this data and try it again later, because it most likely will fail again and it is most likely out of sequence anyway.