Attempt to rewrite device entry in the child table if sending fails
Sometimes the coordinator stops setting the PENDING flag before sending frames to sleepy end devices.
Rewriting the device entry in the child table seems to fix this, at least temporarily.
If sending a frame fails, try to rewrite the device's entry and try again
(if there is already an entry, and only if the device is listed as a sleepy end device, aka CHILD_RFD)