fix: Fix no `checkinInterval` after restart if device has no `genPollCtl` cluster (#1119)
When modernExtend.quirkCheckinInterval() got added, this used to work. But now the restoring of the value from database was behinda check on inputClusters.genPollCtrl.
This modernExtend was added for devices with a fixed checkin interval (e.g. lumi devices and there near frequent genBasic messages) but that do not expose genPollCtrl.