fix: Remove `heimanSpecificFormaldehydeMeasurement` in favour of `msFormaldehyde` (#1166)
* Update cluster.ts | 

Because this is a standard ZCL cluster, not some specific heiman cluster. Due to the current implementation, when configuring reports of any device working with this standard cluster, z2m sends the manufacturer's code in a z2m packet, and the devices reject such a setting. This needs to be fixed. If heiman has such an implementation, then please implement it in the converter of a specific device from heiman.

* Update tstype.ts

msFormaldehyde

* Update tstype.ts

Because this is a standard ZCL cluster, not some specific cluster. Due to the current implementation, when configuring reports of any device working with this standard cluster, z2m sends the manufacturer's code in a z2m packet, and the devices reject such a setting. This needs to be fixed. If Hyman has such an implementation, then please implement it in the converter of a specific device from Hyman.

* Update cluster.ts

---------

Co-authored-by: Koen Kanters <koenkanters94@gmail.com>