feat: Add manuSpecificPhilips3 cluster definition.
The Philips Hue ecosystem uses cluster 0xFC01 for "entertainment mode"
packets. This is a manufacturer-specific extension, that allows high-frequency
updates (tested up to 90 *per second*, but typically 10-50) for up to 10 lights.

Unfortunately, `manuSpecificLegrandDevices` already "owns" 0xFC01, making it
impossible to send the required commands to make this work, even with the
necessary patch to enable full control in `zcl_command`:

  https://github.com/Koenkk/zigbee-herdsman-converters/pull/8867

With this change, the `.cluster` field in `zcl_command` can be specified as
`manuSpecificPhilips3`, and will be correctly identified by zigbee-herdsman.

These changes (this PR and the one mentioned above) is sufficient to allow
Entertainment Mode control over Zigbee2MQTT!

The upcoming beta version of [Bifrost](https://github.com/chrivers/bifrost) will
have rudimentary support for this.