Support ZG-102ZM's interview quirk
ZG-102ZM will fail to interview due to the infamous Error: Interview failed because of failed IAS enroll (zoneState didn't change. 

Adding this quirk makes it functional.

I'm also a bit sad that this thing on Aliexpress says it supports tilt/accelerometer but seems to only be for vibrations. The `contact` field doesn't seem to send any data, but that's a problem for another day.